requirements.txt
torch
transformers
diffusers
ftfy
accelerate
Pillow
matplotlib
scipy